pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "offtarget_primer_tool"
version = "0.1.0"
description = "hg38 Off-target + Primer Designer (Lab Edition v2)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "primer3-py>=2.0.2",
    "pyfaidx>=0.7.2.2",
    "pyranges>=0.0.129",
    "streamlit>=1.30.0",
    "matplotlib>=3.8.0",
    "joblib>=1.3.0",
    "openpyxl>=3.1.2",
    "pytest>=8.0.0",
]

[project.scripts]
otp = "otp.pipeline:main"
